Friendbot, comments: moved middleware back into services/horizon/internal
