+++ services/horizon/internal/middleware_logger.go
@@ -9,7 +9,7 @@ import (
 
 	gctx "github.com/goji/context"
 	"github.com/stellar/go/services/horizon/internal/log"
-	"github.com/stellar/go/services/horizon/render"
+	"github.com/stellar/go/services/horizon/internal/render"
 	"github.com/zenazn/goji/web"
 	"github.com/zenazn/goji/web/middleware"
 	"github.com/zenazn/goji/web/mutil"
